fix: resolved syntax errors related to urn, doi, and prism
DOI's explicitly do not provide a namespace URI due to the use of the URN system (see: http://en.wikipedia.org/wiki/Uniform_resource_name#Absence_of_DOI_namespace). However URN's are not a namespace, but a DNS level registry that is not widely addopted.

An attempt to use http://dx.doi.org/ as a DOI namespace failed because turtle does not allow slashes after a prefix. Resolution was to use a dereferencable URI.