Replace `statisticalTest` by `prov:type`
Similarly to what was done to delete the `statisticType` term.

- For inference defines a second `prov:type` attributes to specify the
type of test in use. In this example for a one-sided we have:
As suggested in NIDASH-DDWG call 7/04/2014.

- For each statistical map defines a second `prov:type` attributes to
specify the type of statistic in use. In this example for t-test we
have:
prov:type = 'nidm:tStatisticalMap'
- For each contrast defines a second `prov:type` attributes to specify
the type of statistic in use. In this example for t-test we have:
`prov:type = 'nidm:tContrast'`